# backend/docs/swagger.yaml
basePath: /
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  handlers.CreateUserRequest:
    description: Datos para crear un nuevo usuario
    properties:
      email:
        example: juan.perez@example.com
        type: string
      name:
        example: Juan Pérez
        type: string
      password:
        example: contraseña123
        type: string
      roleId:
        example: 1
        type: integer
    type: object
...
  handlers.UpdateUserRequest:
    description: Datos para actualizar un usuario existente
    properties:
      email:
        example: juan.perez.updated@example.com
        type: string
      name:
        example: Juan Pérez Actualizado
        type: string
      password:
        example: nuevacontraseña123
        type: string
      roleId:
        example: 2
        type: integer
    type: object
...
  models.User:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      password:
        type: string
      roleId:
        type: integer
      status:
        type: boolean
      updatedAt:
        type: string
    type: object
host: localhost:4000
info:
  contact:
    email: support@example.com
    name: API Support
  description: API para gestión de solicitudes de crédito y evaluación de riesgo crediticio.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Credit Risk Assessment API
  version: "1.0"
paths:
...
  /users:
    get:
      consumes:
      - application/json
      description: Retorna una lista de todos los usuarios del sistema
      produces:
      - application/json
      responses:
        "200":
          description: Lista de usuarios
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "500":
          description: Error interno del servidor
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Obtener todos los usuarios
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Crea un nuevo usuario en el sistema
      parameters:
      - description: Datos del usuario a crear
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Usuario creado exitosamente
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Solicitud inválida
          schema:
            type: string
        "409":
          description: El usuario ya existe
          schema:
            type: string
        "500":
          description: Error interno del servidor
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Crear un nuevo usuario
      tags:
      - Users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Elimina un usuario del sistema
      parameters:
      - description: ID del usuario
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Usuario eliminado exitosamente
        "400":
          description: ID inválido
          schema:
            type: string
        "404":
          description: Usuario no encontrado
          schema:
            type: string
        "500":
          description: Error interno del servidor
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Eliminar un usuario
      tags:
      - Users
    get:
      consumes:
      - application/json
      description: Retorna los detalles de un usuario específico
      parameters:
      - description: ID del usuario
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Usuario encontrado
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: ID inválido
          schema:
            type: string
        "404":
          description: Usuario no encontrado
          schema:
            type: string
        "500":
          description: Error interno del servidor
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Obtener un usuario por ID
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Actualiza los datos de un usuario existente
      parameters:
      - description: ID del usuario
        in: path
        name: id
        required: true
        type: integer
      - description: Datos actualizados del usuario
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Usuario actualizado exitosamente
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Solicitud inválida
          schema:
            type: string
        "404":
          description: Usuario no encontrado
          schema:
            type: string
        "409":
          description: El email ya existe
          schema:
            type: string
        "500":
          description: Error interno del servidor
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Actualizar un usuario
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    description: 'Ingresa el token JWT con el prefijo Bearer. Ejemplo: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"